Update UserType, Animation, CreateAccount

+++ app/src/main/java/com/nguyenanhtrung/schoolmanagement/data/remote/datasource/userid/UserIdRemoteDataSource.kt
@@ -4,5 +4,5 @@ import com.nguyenanhtrung.schoolmanagement.data.local.model.Resource
 
 interface UserIdRemoteDataSource {
 
-    suspend fun getMaxUserId(): Resource<Int>
+    suspend fun getMaxUserId(): Resource<Long>
 }
